DRY typings, extract speech synthesis into module.
