+++ koala/shared-types.ts
@@ -1,0 +1,7 @@
+/** TYPES ONLY IN THIS FILE PLEASE */
+
+export type Gender = "M" | "F" | "N";
+export type LangCode = "ko" | "es" | "it" | "fr";
+export type LessonType = "listening" | "speaking";
+export type QuizResult = "error" | "fail" | "pass";
+export type YesNo = "yes" | "no";
